Kien

+++ ArtWorkSharingAPI/Controllers/InteractController.cs
@@ -1,5 +1,6 @@
 ﻿using AWS_BusinessObjects.Entities;
 using AWS_Services.Interface;
+using AWS_Services.Services;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ArtWorkSharingAPI.Controllers
@@ -24,31 +25,71 @@ namespace ArtWorkSharingAPI.Controllers
         public IActionResult GetById(Guid id)
         {
             var interact = interactService.GetById(id);
-            return Ok(interact);
+            if (interact == null)
+            {
+                return NotFound($"Không tìm thấy tương tác của bạn!, Id: {id}");
+            } 
+            else
+            {
+                return Ok(interact);
+            }          
         }
         [HttpPost("Add")]
         public IActionResult Add(Interact interact)
         {
-            interactService.Add(interact);
-            return Ok("Thêm thành công");
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            else
+            {
+                interactService.Add(interact);
+                return Ok("Thêm thành công");
+            }
         }
         [HttpDelete("Delete")]
         public IActionResult Delete(Guid id)
         {
-            interactService.Delete(id);
-            return Ok("Xóa thành công");
+            if (ModelState.IsValid)
+            {
+                var interactCheck = interactService.GetById(id);
+                if (interactCheck == null)
+                {
+                    ModelState.AddModelError($"Id", $"Không tìm thấy tương tác của bạn!, Id: {id}");
+                    return NotFound(ModelState);
+                }
+                else
+                {
+                    interactService.Delete(id);
+                    return Ok("Xóa thành công");
+                }
+            }
+            else
+            {
+                return BadRequest($"Số Lỗi: {ModelState.ErrorCount}, Lỗi: {ModelState}");
+            }
         }
         [HttpPut("Update")]
         public IActionResult Update(Interact interact)
         {
-            var Interact = interactService.GetById(interact.Id);
-
-            if (Interact != null)
+            if (ModelState.IsValid)
             {
-                interactService.Update(interact);
-                return Ok("Cập nhật thành công");
+                var interactCheck = interactService.GetById(interact.Id);
+                if (interactCheck == null)
+                {
+                    ModelState.AddModelError($"Id", $"Không tìm thấy tương tác của bạn!, Id: {interact.Id}");
+                    return NotFound(ModelState);
+                }
+                else
+                {
+                    interactService.Update(interact);
+                    return Ok("Cập nhật thành công");
+                }
             }
-            return NotFound();
+            else
+            {
+                return BadRequest($"Số Lỗi: {ModelState.ErrorCount}, Lỗi: {ModelState}");
+            }
         }
     }
 }
