+++ ArtWorkSharingAPI/Controllers/PackageDetailController.cs
@@ -1,5 +1,6 @@
 ﻿using AWS_BusinessObjects.Entities;
 using AWS_Services.Interface;
+using AWS_Services.Services;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ArtWorkSharingAPI.Controllers
@@ -24,31 +25,74 @@ namespace ArtWorkSharingAPI.Controllers
         public IActionResult GetById(Guid id)
         {
             var packageDetail = packageDetailsService.GetById(id);
-            return Ok(packageDetail);
+            if (packageDetail == null)
+            {
+                return NotFound($"Không tìm thấy chi tiết gói của bạn!, Id: {id}");
+            }
+            else
+            {
+                return Ok(packageDetail);
+            }       
+            
         }
         [HttpPost("Add")]
         public IActionResult Add(PackageDetail packageDetail)
         {
-            packageDetailsService.Add(packageDetail);
-            return Ok("Thêm thành công");
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            else
+            {
+                packageDetailsService.Add(packageDetail);
+                return Ok("Thêm thành công");
+            }
+            
         }
         [HttpDelete("Delete")]
         public IActionResult Delete(Guid id)
         {
-            packageDetailsService.Delete(id);
-            return Ok("Xóa thành công");
+            if (ModelState.IsValid)
+            {
+                var packageDetailCheck = packageDetailsService.GetById(id);
+                if (packageDetailCheck == null)
+                {
+                    ModelState.AddModelError($"Id", $"Không tìm thấy chi tiết gói của bạn!, Id: {id}");
+                    return NotFound(ModelState);
+                }
+                else
+                {
+                    packageDetailsService.Delete(id);
+                    return Ok("Xóa thành công");
+                }
+            }
+            else
+            {
+                return BadRequest($"Số Lỗi: {ModelState.ErrorCount}, Lỗi: {ModelState}");
+            }
+            
         }
         [HttpPut("Update")]
         public IActionResult Update(PackageDetail packageDetail)
         {
-            var PackageDetail = packageDetailsService.GetById(packageDetail.Id);
-
-            if (PackageDetail != null)
+            if (ModelState.IsValid)
             {
-                packageDetailsService.Update(packageDetail);
-                return Ok("Cập nhật thành công");
+                var packageDetailCheck = packageDetailsService.GetById(packageDetail.Id);
+                if (packageDetailCheck == null)
+                {
+                    ModelState.AddModelError($"Id", $"Không tìm thấy chi tiết gói của bạn!, Id: {packageDetail.Id}");
+                    return NotFound(ModelState);
+                }
+                else
+                {
+                    packageDetailsService.Update(packageDetail);
+                    return Ok("Cập nhật thành công");
+                }
             }
-            return NotFound();
+            else
+            {
+                return BadRequest($"Số Lỗi: {ModelState.ErrorCount}, Lỗi: {ModelState}");
+            }
         }
     }
 }
